[GLES] Move the Hook class declaration to a header file
